cbfstool: Support 64bit addresses for flat images
SELF has the fields wired up for 64bit, but adding flat images cuts the
upper half.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/80576
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ util/cbfstool/cbfs-mkpayload.c
@@ -230,8 +230,8 @@ out:
 
 int parse_flat_binary_to_payload(const struct buffer *input,
 				 struct buffer *output,
-				 uint32_t loadaddress,
-				 uint32_t entrypoint,
+				 uint64_t loadaddress,
+				 uint64_t entrypoint,
 				 enum cbfs_compression algo)
 {
 	comp_func_ptr compress;
